slackbot/requirements.txt
pymongo
sqlalchemy
psycopg2-binary
requests
pandas
slackclient